+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+- Fix problems when working with shared_preferences.
+  - Fix `WidgetsFlutterBinding.ensureInitialized()` issue.
+- Update example to use shared_preferences with locale.
+
 ## 1.0.0
 
 - Initial release.